Hot-Sync
Hot Syncing is the use of urban Wifi router device planning and synced connectivity processing to compute a large area or sector of Wifi to operate as one segment of connectivity. Typical to a network inside one building, this sort of syncing is designed to allow for an entire urban area or large city sector to work as one.